git-svn-id: svn://svn.code.sf.net/p/win32-framework/code@2783 40094c20-7226-0410-9e7a-ca24b3fb783e

+++ samples/CommonDialogs/src/Mainfrm.h
@@ -37,7 +37,7 @@ CMainFrame : public CFrame                                                  /*
 
         AboutBox& GetAboutBox() { return m_aboutBox; }
         void    EmptyMRUList();
-        size_t  GetMRUSize() { return GetMRUEntries().size();}
+        UINT   GetMRUSize() {return static_cast<UINT>(GetMRUEntries().size());}
         COLORREF GetSBBkColor()
                     { return m_colorChoice.GetTableColor(SBBg);}
         void    SetSBBkColor(COLORREF clr)
